requirements.txt
requests
datetime
python-dateutil
boto3
bs4
genanki